% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-bcea-plotting-fns.R
\name{my.plot.bcea}
\alias{my.plot.bcea}
\title{my.plot.bcea}
\usage{
my.plot.bcea(dat1, dat2 = NA, dat3 = NA, dat4 = NA, wtp = 20000,
  intlabels, contour = FALSE, LEVELS = 0.95, XLIM = c(-20, 20),
  YLIM = c(-1e+05, 50000), wtpNEG = NA, SCALEdays = TRUE,
  SCALEcosts = TRUE, labelLong = TRUE, TITLE = "", N = 1)
}
\arguments{
\item{dat1}{(e,c) pairs}

\item{dat2}{(e,c) pairs}

\item{dat3}{(e,c) pairs}

\item{dat4}{(e,c) pairs}

\item{wtp}{Willingness-to-pay, may be a vector i.e. multiple lines}

\item{intlabels}{Intervention labels to use in the legend}

\item{contour}{Include ellipses?}

\item{LEVELS}{0-1 confidence intervals for ellipse(s)}

\item{SCALEdays}{x-axis QALYs in days or years}

\item{SCALEcosts}{should cost be per person}

\item{labelLong}{Full point description or not}

\item{N}{original sample size}
}
\value{
plot
}
\description{
Cost-effectiveness plane plot.
Borrowed ideas from the \code{BCEA} package.
}
