% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my-bcea-plotting-fns.R
\name{my.plot.ceac}
\alias{my.plot.ceac}
\title{My Version of the Cost-Effectiveness Acceptability Curve Plot}
\usage{
my.plot.ceac(dat1, dat2, dat3, dat4 = NA, intlabels, wtpNEG = NA,
  labelLong = TRUE, TITLE = "", SCALEdays = TRUE, SCALEcosts = TRUE,
  N = 1, CI = FALSE)
}
\arguments{
\item{dat1}{Output from IDEAdectree.simple}

\item{dat2}{Output from IDEAdectree.simple}

\item{dat3}{Output from IDEAdectree.simple}

\item{dat4}{Output from IDEAdectree.simple}

\item{intlabels}{Intervention labels to use in the legend}

\item{wtpNEG}{If no health detriment is ok. Asymptotic as wtp tends to infinity.}

\item{labelLong}{(TRUE)}

\item{SCALEdays}{(TRUE)}

\item{SCALEcosts}{(TRUE)}

\item{N}{Original sample size (number of patients)}

\item{CI}{Use Binomial 95% confidence intervals, capture some model uncertainty due to simulation size}
}
\value{
plot
}
\description{
Borrowed ideas from the \code{BCEA} package.
}
\seealso{
\link{\code{IDEAdectree.simple}}
}
