% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ceac_plot.R
\name{my_ceac.plot}
\alias{my_ceac.plot}
\title{Plot a Cost-Effectiveness Acceptability Curve}
\usage{
my_ceac.plot(screen.bcea, new_window = FALSE, scenario = NA)
}
\arguments{
\item{screen.bcea}{A BCEA object from the BCEA package}

\item{new_window}{default: FALSE}

\item{scenario}{Specific scenario numbers to plot. default: NA}
}
\value{

}
\description{
The function that comes with the \code{BCEA} package
is a bit basic and hard to modify.
}
\details{
TODO: ggplot version
}
\examples{

}
