% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_ceplane_plot (1).R, R/my_ceplane_plot.R
\name{my_ceplane.plot}
\alias{my_ceplane.plot}
\title{my_ceplane.plot}
\usage{
my_ceplane.plot(
  he,
  comparison = NULL,
  wtp = 25000,
  pos = c(1, 1),
  size = NULL,
  graph = c("base", "ggplot2"),
  xlim = NULL,
  ylim = NULL,
  withpoints = FALSE,
  ...
)

my_ceplane.plot(
  he,
  comparison = NULL,
  wtp = 25000,
  pos = c(1, 1),
  size = NULL,
  graph = c("base", "ggplot2"),
  xlim = NULL,
  ylim = NULL,
  withpoints = FALSE,
  ...
)
}
\arguments{
\item{...}{}
}
\value{



}
\description{
my_ceplane.plot

my_ceplane.plot
}
