% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s_analysis_to_tornado_plot_data.R
\name{s_analysis_to_tornado_plot_data}
\alias{s_analysis_to_tornado_plot_data}
\title{Convert Sensitivity Analysis Output Data to Tornado Plot Input Data}
\usage{
s_analysis_to_tornado_plot_data(s_analysis, baseline_input = NA)
}
\arguments{
\item{s_analysis}{model.frame object}

\item{baseline_input}{Named vector of baseline parameter values}
}
\value{
Data frame of class tornado
}
\description{
A preprocessing step is required to use \code{ggplot_tornado}.
}
\details{
This can be thougth of as going from a multi-way design to a one-way design.

The output data from a multivariate sensitivity analysis are usually in the form
of a grid of input parameter values and a column of associated output values.

\tabular{rrrr}{
  \cr
  variable1 \tab variable2 \tab variable3 \tab output\cr
  0 \tab 0 \tab 0 \tab 1\cr
  1 \tab 0 \tab 0 \tab 2\cr
  0 \tab 1 \tab 0 \tab 5\cr
}

For a tornado plot we only want to keep the maximum and minimum output values
for each parameter when all others are set at baseline (one-way analysis).

We also want to record whether low parameter values give low output values
or vice-versa.
}
\examples{

s_analysis <- data.frame(output = c(10,1,11,5,3),
                         sens = c(2,2,3,0,2),
                         spec = c(1,4,2,2,2))

s_analysis <- model.frame(formula = output ~ sens + spec,
                          data = s_analysis)

s_analysis_to_tornado_plot_data(s_analysis)

}
\seealso{
\code{\link{ggplot_tornado}}
}
