% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tornado_plot.R
\name{tornado_plot}
\alias{tornado_plot}
\title{A Simple Tornado Plot in Base R}
\usage{
tornado_plot(
  cf,
  se,
  df = NULL,
  level = 0.95,
  parm = NULL,
  labels = FALSE,
  xlab = "Coefficient confidence intervals",
  ylab = "",
  xlim = NULL,
  ylim = NULL,
  las = 1,
  lwd = 1,
  lty = c(1, 2),
  pch = 19,
  col = 1,
  length = 0,
  angle = 30,
  code = 3,
  ...
)
}
\arguments{
\item{cf}{Central values}

\item{se}{Standard errors}

\item{df}{df}

\item{level}{level}

\item{parm}{parm}

\item{labels}{labels}

\item{xlab}{xlab}

\item{ylab}{ylab}

\item{xlim}{xlim}

\item{ylim}{ylim}

\item{las}{las}

\item{lwd}{lwd}

\item{lty}{lty}

\item{pch}{pch}

\item{col}{col}

\item{length}{length}

\item{angle}{angle}

\item{code}{code}

\item{...}{...}
}
\value{
plot
}
\description{
Provide middle value and standard errors.
Assumes symmetrical.
}
\details{
https://www.r-statistics.com/2010/07/visualization-of-regression-coefficients-in-r/
}
\examples{
tornado_plot(cf = c(a=2, b=3), se = c(1,2), labels = TRUE)

}
