% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/basic_tornado_plot.R
\name{tornado_plot}
\alias{tornado_plot}
\title{A Simple Tornado Plot}
\usage{
tornado_plot(cf, se, df = NULL, level = 0.95, parm = NULL,
  labels = TRUE, xlab = "Coefficient confidence intervals", ylab = "",
  xlim = NULL, ylim = NULL, las = 1, lwd = 1, lty = c(1, 2),
  pch = 19, col = 1, length = 0, angle = 30, code = 3, ...)
}
\arguments{
\item{cf}{Central values}

\item{se}{Standard errors}

\item{...}{}
}
\description{
https://www.r-statistics.com/2010/07/visualization-of-regression-coefficients-in-r/
}
\examples{
tornado_plot(cf = c(2,3), se = c(1,2))

}
